+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Ephemeral nodes can be hidden from DAG by setting `show_ephemeral_models: False` in project's `airflow.yml`.
+
 ## [0.21.0] - 2022-02-11
 
 This version brings compatibility with `dbt 1.0`.